feat(events): extract events ops to event controller
